Balance patch: More recoil to the basic gun and the range of the aliens guns has decreased.

+++ Assets/Scripts/Ammo.cs
@@ -7,7 +7,7 @@ public class Ammo : MonoBehaviour
 
     public int damage;
     public Rigidbody2D rb;
-    public int destroytimer;
+    public float destroytimer;
 
     // Use this for initialization
     void Start()
